+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Removed
+- Modules: 
+  - `movements-top-bar`
+
 ## [1.13.0-rc.2] - 2021-02-15
 ### Changed
 - Refactor function addBalance
-- Modules: 
-  - `movements-top-bar`
 
 ## [1.13.0-rc.1] - 2021-02-09
 #### Added